+++ django_mysql/models/fields/sets.py
@@ -72,7 +72,14 @@ class SetFieldMixin(object):
 
     def deconstruct(self):
         name, path, args, kwargs = super(SetFieldMixin, self).deconstruct()
-        path = 'django_mysql.models.%s' % self.__class__.__name__
+
+        bad_paths = (
+            'django_mysql.models.fields.sets.' + self.__class__.__name__,
+            'django_mysql.models.fields.' + self.__class__.__name__
+        )
+        if path in bad_paths:
+            path = 'django_mysql.models.' + self.__class__.__name__
+
         args.insert(0, self.base_field)
         kwargs['size'] = self.size
         return name, path, args, kwargs