+++ HISTORY.rst
@@ -9,6 +9,9 @@ History
 * Pending release
 * ``SmartChunkedIterator`` with ``report_progress=True`` correctly reports
   'lowest pk so far' when iterating in reverse.
+* Fix broken import paths during ``deconstruct()`` for subclasses of all
+  fields: ``ListCharField``, ``ListTextField``, ``SetCharField``,
+  ``SetTextField``, ``SizedBinaryField`` and ``SizedTextField``
 
 0.2.3 (2015-10-12)
 ------------------